refactor(scatter-plot): group helpers into feature subdirectories
Organize the flat scatter-plot/ helper files into feature folders so the directory is easier to navigate. No behavior change — only file moves (via git mv, history preserved) plus relative-import path updates.

New layout (entry, config, styles, perf, integration tests, and the webgl/ subtree stay at the root):

- tooltips/            protein-tooltip(+styles,+helpers), protspace-tips(+styles), tooltip-position, tooltip-height-estimate

- duplicate-stacks/    duplicate-badges-canvas-renderer, duplicate-stack-{helpers,overlay-controller,types,viewport}, spiderfy-layer

- interaction/         plot-interaction-controller, quadtree-index

- styling/             visibility-model, style-getters, numeric-recompute-runner

- projection-metadata/ projection-metadata(+styles)
