+++ packages/core/src/components/scatter-plot/duplicate-stacks/duplicate-badges-canvas-renderer.ts
