+++ packages/core/src/components/scatter-plot/duplicate-stacks/duplicate-stack-helpers.test.ts
