+++ packages/core/src/components/scatter-plot/duplicate-stacks/spiderfy-layer.test.ts
