+++ packages/core/src/components/scatter-plot/interaction/plot-interaction-controller.test.ts
