+++ packages/core/src/components/scatter-plot/projection-metadata/projection-metadata.styles.ts
