+++ packages/core/src/components/scatter-plot/styling/style-getters.test.ts
