+++ packages/core/src/components/scatter-plot/styling/visibility-model.test.ts
