+++ packages/core/src/components/scatter-plot/tooltips/protein-tooltip-helpers.test.ts
