+++ packages/core/src/components/scatter-plot/tooltips/protein-tooltip.styles.ts
