+++ packages/core/src/components/scatter-plot/tooltips/protein-tooltip.ts
@@ -1,6 +1,6 @@
 import { LitElement, html, type TemplateResult } from 'lit';
 import { property } from 'lit/decorators.js';
-import { customElement } from '../../utils/safe-custom-element';
+import { customElement } from '../../../utils/safe-custom-element';
 import { toDisplayValue, toInternalValue } from '@protspace/utils';
 import type { AnnotationBlock, NumericAnnotationType, TooltipView } from '@protspace/utils';
 import { proteinTooltipStyles } from './protein-tooltip.styles';
