+++ packages/core/src/components/scatter-plot/tooltips/protspace-tips.styles.ts
