+++ packages/core/src/components/scatter-plot/tooltips/protspace-tips.ts
@@ -1,6 +1,6 @@
 import { LitElement, html, nothing } from 'lit';
 import { property, state } from 'lit/decorators.js';
-import { customElement } from '../../utils/safe-custom-element';
+import { customElement } from '../../../utils/safe-custom-element';
 import { isMacOrIos } from '@protspace/utils';
 import { protspaceTipsStyles } from './protspace-tips.styles';
