+++ packages/core/src/components/scatter-plot/tooltips/tooltip-height-estimate.ts
